+++ src/Components/Attendees.js
@@ -1,3 +1,11 @@
-export default function Attendees() {
-  return;
+import Attendee from "./Attendee"
+
+export default function Attendees({event, attendee, updateEventAttendance}) {
+  return (
+    <Attendee 
+    attendee={attendee} 
+    event={event} 
+    updateEventAttendance={updateEventAttendance}
+    />
+  );
 }
